Merge branch 'master' into release/2.2.0

+++ python/opae.admin/opae/admin/tools/opaevfio.py
@@ -37,7 +37,7 @@ import subprocess
 import sys
 import time
 
-OPAEVFIO_VERSION = '1.0.2'
+OPAEVFIO_VERSION = '1.0.3'
 
 ABBREV_PCI_ADDR_PATTERN = r'([\da-fA-F]{2}):' \
                           r'([\da-fA-F]{2})\.' \
@@ -183,14 +183,33 @@ def initialize_vfio(addr, new_owner, enable_sriov):
         load_driver('vfio-pci')
 
     print('Binding {} to vfio-pci'.format(msg))
-    new_id = '/sys/bus/pci/drivers/vfio-pci/new_id'
-    try:
-        with open(new_id, 'w') as outf:
-            outf.write('{} {}'.format(vid_did[0], vid_did[1]))
-    except OSError as exc:
-        if exc.errno != errno.EEXIST:
-            print(exc)
+
+
+    # On Linux kernel >= 3.16, use driver_override to specify the
+    # driver that may bind to the function with the given address.
+    # On older kernels, fall back to new_id which matches the vendor
+    # and device ID to determine whether a driver may bind to a
+    # device, which has the disadvantage that the driver may bind
+    # to multiple functions with the same vendor and device ID.
+    #
+    # https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=782a985d7af26db39e86070d28f987cad21313c0
+    driver_override = '/sys/bus/pci/devices/{}/driver_override'.format(addr)
+    if os.path.exists(driver_override):
+        try:
+            with open(driver_override, 'w') as outf:
+                outf.write('vfio-pci')
+        except OSError as exc:
+            print('Cannot write driver_override to vfio-pci for: {}'.format(msg, exc))
             return
+    else:
+        new_id = '/sys/bus/pci/drivers/vfio-pci/new_id'
+        try:
+            with open(new_id, 'w') as outf:
+                outf.write('{} {}'.format(vid_did[0], vid_did[1]))
+        except OSError as exc:
+            if exc.errno != errno.EEXIST:
+                print(exc)
+                return
 
     time.sleep(0.50)
 
@@ -250,6 +269,19 @@ def release_vfio(addr, new_driver):
         return
 
     print('Releasing {} from vfio-pci'.format(msg))
+
+    # On Linux kernel >= 3.16, clear driver_override before
+    # unbinding vfio-pci driver, which ensures subsequent
+    # binding to another, previously bound driver succeeds.
+    driver_override = '/sys/bus/pci/devices/{}/driver_override'.format(addr)
+    if os.path.exists(driver_override):
+        try:
+            with open(driver_override, 'w') as outf:
+                outf.write('\n')
+        except OSError as exc:
+            print('Cannot clear driver_override for {}: {}'.format(msg, exc))
+            return
+
     unbind_driver(driver, addr)
 
     if new_driver and bind_driver(new_driver, addr):
